address PR feedback and move type declarations to util file
