+++ ui/lib/core/addon/utils/client-count-utils.ts
@@ -7,20 +7,14 @@ import { parseAPITimestamp } from 'core/utils/date-formatters';
 import { compareAsc, getUnixTime, isWithinInterval } from 'date-fns';
 
 import type ClientsVersionHistoryModel from 'vault/vault/models/clients/version-history';
-import type {
-  ActivityMonthBlock,
-  ByMonthClients,
-  ByMonthNewClients,
-  ByNamespaceClients,
-  Counts,
-  EmptyActivityMonthBlock,
-  MountByKey,
-  MountClients,
-  MountNewClients,
-  NamespaceByKey,
-  NamespaceNewClients,
-  NamespaceObject,
-} from 'vault/vault/utils/client-count-utils';
+
+/*
+The client count utils are responsible for serializing the sys/internal/counters/activity API response
+The initial API response shape and serialized types are defined below and used to
+defines the activity model in models/clients/activity.d.ts
+
+To help visualize there are sample responses in ui/tests/helpers/clients.js
+*/
 
 // add new types here
 export const CLIENT_TYPES = [
@@ -152,7 +146,7 @@ export const sortMonthsByTimestamp = (monthsArray: ActivityMonthBlock[] | EmptyA
 
 export const namespaceArrayToObject = (
   monthTotals: ByNamespaceClients[],
-  // technically this arg is the same type as above, just nested
+  // technically this arg (monthNew) is the same type as above, just nested inside monthly new clients
   monthNew: ByMonthClients['new_clients']['namespaces'],
   month: string,
   timestamp: string
@@ -205,4 +199,118 @@ export function hasNamespacesKey(
   obj: ByMonthNewClients | NamespaceNewClients | MountNewClients
 ): obj is ByMonthNewClients {
   return 'namespaces' in obj;
+}
+
+// TYPES RETURNED BY UTILS (serialized)
+
+export interface TotalClients {
+  clients: number;
+  entity_clients: number;
+  non_entity_clients: number;
+  secret_syncs: number;
+  acme_clients: number;
+}
+
+export interface ByNamespaceClients extends TotalClients {
+  label: string;
+  mounts: MountClients[];
+}
+
+export interface MountClients extends TotalClients {
+  label: string;
+}
+
+export interface ByMonthClients extends TotalClients {
+  month: string;
+  timestamp: string;
+  namespaces: ByNamespaceClients[];
+  namespaces_by_key: { [key: string]: NamespaceByKey };
+  new_clients: ByMonthNewClients;
+}
+
+// interface EmptyByMonthClients {
+//   month: string;
+//   timestamp: string;
+//   namespaces: [];
+//   namespaces_by_key: Record<string, never>;
+//   new_clients: {
+//     month: string;
+//     timestamp: string;
+//     namespaces: [];
+//   };
+// }
+export interface ByMonthNewClients extends TotalClients {
+  month: string;
+  timestamp: string;
+  namespaces: ByNamespaceClients[];
+}
+
+export interface NamespaceByKey extends TotalClients {
+  month: string;
+  timestamp: string;
+  mounts_by_key: { [key: string]: MountByKey };
+  new_clients: NamespaceNewClients;
+}
+
+export interface NamespaceNewClients extends TotalClients {
+  month: string;
+  label: string;
+  mounts: MountClients[];
+}
+
+export interface MountByKey extends TotalClients {
+  month: string;
+  timestamp: string;
+  label: string;
+  new_clients: MountNewClients;
+}
+
+export interface MountNewClients extends TotalClients {
+  month: string;
+  label: string;
+}
+
+// API RESPONSE SHAPE (prior to serialization)
+
+interface SysInternalCountersActivityResponse {
+  start_time: string;
+  end_time: string;
+  total: Counts;
+  by_namespace: NamespaceObject[];
+  months: ActivityMonthBlock[];
+}
+
+export interface NamespaceObject {
+  namespace_id: string;
+  namespace_path: string;
+  counts: Counts;
+  mounts: { mount_path: string; counts: Counts }[];
+}
+
+export interface ActivityMonthBlock {
+  timestamp: string; // YYYY-MM-01T00:00:00Z (always the first day of the month)
+  counts: Counts;
+  namespaces: NamespaceObject[];
+  new_clients: {
+    counts: Counts;
+    namespaces: NamespaceObject[];
+    timestamp: string;
+  };
+}
+
+export interface EmptyActivityMonthBlock {
+  timestamp: string; // YYYY-MM-01T00:00:00Z (always the first day of the month)
+  counts: null;
+  namespaces: null;
+  new_clients: null;
+}
+
+export interface Counts {
+  acme_clients: number;
+  clients: number;
+  distinct_entities: number;
+  entity_clients: number;
+  non_entity_clients: number;
+  non_entity_tokens: number;
+  secret_syncs: number;
 }
