+++ ui/types/vault/models/clients/activity.d.ts
@@ -5,7 +5,8 @@
 
 import type { Model } from 'vault/app-types';
 
-import type { ByMonthClients, ByNamespaceClients, TotalClients } from 'vault/vault/utils/client-count-utils';
+import type { ByMonthClients, ByNamespaceClients, TotalClients } from 'core/utils/client-count-utils';
+
 export default interface ClientsActivityModel extends Model {
   byMonth: ByMonthClients[];
   byNamespace: ByNamespaceClients[];
